gpu 	modified:   ../dftbtorch/dftbcalculator.py 	modified:   ../dftbtorch/geninterpskf.py 	modified:   ../dftbtorch/initparams.py 	modified:   ../dftbtorch/interpolator.py 	modified:   ../dftbtorch/matht.py 	modified:   ../dftbtorch/mbd.py 	modified:   ../dftbtorch/mixer.py 	modified:   ../dftbtorch/sk.py 	modified:   ../ml/interface.py 	modified:   dftb_in.mlCompressionR 	modified:   dftbt.py

+++ dftbtorch/geninterpskf.py
@@ -565,16 +565,16 @@ class SkInterpolator:
                 spe[icount] = fp_line2_[3]
                 uhubb[icount, :] = fp_line2_[4:7]
                 occ_skf[icount, :] = fp_line2_[7:10]
-                data = np.fromfile(fp, dtype=float, count=20, sep=' ')
+                data = np.fromfile(fp, count=20, sep=' ')
                 mass_rcut[icount, :] = t.from_numpy(data)
-                data = np.fromfile(fp, dtype=float, count=nitem, sep=' ')
+                data = np.fromfile(fp, count=nitem, sep=' ')
                 data.shape = (int(ngridpoint[icount]), 20)
                 integrals.append(data)
                 self.para['rest'].append(fp.read())
             else:
-                data = np.fromfile(fp, dtype=float, count=20, sep=' ')
+                data = np.fromfile(fp, count=20, sep=' ')
                 mass_rcut[icount, :] = t.from_numpy(data)
-                data = np.fromfile(fp, dtype=float, count=nitem, sep=' ')
+                data = np.fromfile(fp, count=nitem, sep=' ')
                 data.shape = (int(ngridpoint[icount]), 20)
                 integrals.append(data)
                 self.para['rest'].append(fp.read())
